# Conflicts:
#	packages/google_maps_flutter/google_maps_flutter_ios/CHANGELOG.md
#	packages/google_maps_flutter/google_maps_flutter_ios/example/ios14/ios/Runner.xcodeproj/project.pbxproj
#	packages/google_maps_flutter/google_maps_flutter_ios/ios/Classes/GoogleMapMarkerController.h
#	packages/google_maps_flutter/google_maps_flutter_ios/pubspec.yaml

+++ packages/pigeon/platform_tests/test_plugin/android/src/test/kotlin/com/example/test_plugin/PrimitiveTest.kt
@@ -21,7 +21,7 @@ class PrimitiveTest {
     val binaryMessenger = mockk<BinaryMessenger>(relaxed = true)
     val api = mockk<PrimitiveHostApi>(relaxed = true)
 
-    val input = 1
+    val input = 1L
 
     val channelName = "dev.flutter.pigeon.pigeon_integration_tests.PrimitiveHostApi.anInt"
     val handlerSlot = slot<BinaryMessenger.BinaryMessageHandler>()
@@ -38,7 +38,7 @@ class PrimitiveTest {
       it?.rewind()
       @Suppress("UNCHECKED_CAST") val wrapped = codec.decodeMessage(it) as List<Any>?
       assertNotNull(wrapped)
-      wrapped?.let { assertEquals(input.toLong(), wrapped[0]) }
+      wrapped?.let { assertEquals(input, wrapped[0]) }
     }
 
     verify { binaryMessenger.setMessageHandler(channelName, handlerSlot.captured) }
